requirements.txt
django
lxml
requests
pillow
djangorestframework
markdown
django-filter
django-cors-headers
factory-boy